Update README.md to remove auth.json after use
The auth.json file is created with a Github username and password to be used to authenticate requests from Composer back to Github.  The username and password is recommended to use a personal access token to access private repos.  The auth.json file is created at the root of the repo and if your Github Actions job eventually deploys the contents of this repo to a web accessible server, it will deploy the auth.json file too based on how the example is written.  This update adds a step to remove the auth.json file from the repo after it is used by Composer, preventing sensitive credentials from being deployed in plain text to a potentially public location.